requirements.txt
requests
beautifulsoup4
lxml
pytest
Pillow 
tqdm